chore(appliance): split otel collector and agent (#63989)
Relates to

https://linear.app/sourcegraph/issue/REL-81/service-definition-otel-collector
but does not close it.

Some of the StandardConfig features (notably PodTemplateConfig) assume
that there is only one pod template "thing" per service. The otel
service contains an agent daemonset and a collector deployment, which
have quite different characteristics. We're less likely to paint
ourselves into a corner if we split these services at the config level.