authz: Compute providers on the fly
Previously, we would store authz providers globally and refresh them every now and then.
However, creating the providers is fairly cheap (1.3ms in a local trace) so we should not keep them in memory and remember to not forget to start the watcher routine.

This will help for multi-tenant Sourcegraph in that providers are now computed for the context in question, and not held globally. Keeping potentially 100k authz providers in memory will not scale.

Test plan:

Still works, local Jaeger traces are quite acceptable.