feat/enterpriseportal: support DevOnly in list subscriptions (#63837)
We need to ensure prod data does not end up in the dev Enterprise Portal
instance once we start syncing. This change adds support for the DevOnly
option in ListSubscriptions and strictly enforces it there (i.e. dev
subscriptions cannot end up in the prod instance either).

I've tried to minimize the changes elsewhere to reduce the impact of
this change for now.

## Test plan

Tests