Bump openjdk-11 version in blobstore image (#64047)
<!-- PR description tips:
https://www.notion.so/sourcegraph/Write-a-good-pull-request-description-610a7fd3e613496eb76f450db5a49b6e
-->
Update the version of openjdk-11 we use in the blobstore image.

We updated the blobstore image earlier today to try and fix this issue,
but were thwarted by an old version pin. This has now been
[removed](https://github.com/sourcegraph/sourcegraph/pull/64045), so
updating packages fully resolves the issue.

This PR branch is not based off `main` as backporting package changes
often results in merge conflicts. It was generated by checking our
`5.5.x` and running `sg wolfi lock blobstore`.

## Test plan

- CI
- Run image locally

<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->