cody-gateway: handle missing Google response (#63895)
- Logs a warning instead of returning an error when the Google response
is missing
- This prevents the API from returning an error when the Google response
is empty, which can happen when Google is not happy with the question
due to safety issues. We will only log a decoder error as error.

<!-- PR description tips:
https://www.notion.so/sourcegraph/Write-a-good-pull-request-description-610a7fd3e613496eb76f450db5a49b6e
-->

## Test plan

<!-- REQUIRED; info at
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles
-->

only the log level was updated for empty responses. the function
behavior was not changed.

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->

cody-gateway: log missing Google response as warning