chore: Remove redis init side-effect of app.NewHandler
This current code sets a global variable which couples the session package and the app package and the session package cannot work when the NewHandler method wasn't called yet.

This also used to wait for redis implicitly, without any indication of doing that.

To be more explicit about that, we now do that in the main function of frontend instead, and create the session store where required on the fly.

Test plan: Login still works, integration/E2E tests are passing.