auth: Fix Found page response when redis is down
At some point, we decided to serve a 500 error instead of a 302 FOUND which redirects to the sign in page when redis is down. This usually happens during a rollout because redis is usually also rolled out at the same time.
When that happens, users would previously get redirected to the signin page, and their session is cleared.

So we decided to instead send a 500 status code, and let them explicitly retry.

However, that didn't stop the execution chain before, so we would send the 500 status code, but then eventually get to the app handler that calls http.Redirect with 302 FOUND (which as a side-effect renders a `Found.` link as an HTML page for old browsers that don't support 302 natively). Since the http status header has already been sent, it doesn't change the status code to 302 and instead returns a 500 response with the HTML `Found.` link.

This is the strange error page we've been seeing during rollouts a lot of times and never knew where it's from.

Test plan: Locally stopped redis while the instance was running and no longer saw the `Found.` message and instead now get a proper error message.