feat(appliance): healthchecker manages ingress-facing service (#64043)
**Appliance points ingress-facing service to itself by default**

Not frontend.

**feat(appliance): healthchecker manages ingress-facing service**

Add a new background goroutine to the appliance. It does nothing until a
"begin" channel closes. The idea is that another part of the appliance
will close this channel if the configmap state is set to a post-install
value (or on startup if this is already the case when an appliance
boots).

After this barrier is lifted, the healtchecker periodically checks the
readiness (using k8s conditions) of each pod returned by the frontend
deployment's label selector. If even a single pod is ready, it ensures
that the service points to frontend. Otherwise, it waits for a grace
period, checks again, and if downtime persists, it points the service to
the appliance.

This should cover the following cases:
- The service is pointed to frontend after the admin clicks "go" after
  an initial successful install.
- The service is pointed to appliance after frontend downtime that
  exceeds the grace period.
- The service is promptly pointed to frontend after downtime ends.