fix(search_jobs): report as processing even
This fixes a bug in our status reporting where we would report a job as
"queued" although it should be reported as "processing". In some cases
this led to the state flip-flopping several times between "queued" and
"processing".

This happens, for example, if a worker has just finished a job (and set
it to completed) but hasn't yet dequeued a new job. In this brief period
we reported the entire search job as queued.

Test plan:
- updated unit test
- manual testing