chore: Bubble precise Usages instead of Locations (#64118)
This patch changes the code so that the core APIs we need
for precise usagesForSymbol actually return (Upload)Usages instead
of just (Upload)Locations, so that we have the usage kind information
as well as the symbol names available. I will be handling the GraphQL
layer for converting the Usages to appropriate Resolvers in a follow-up PR.

## Test plan

Modified a bunch of existing mock tests. 🤷🏽