fix(search_jobs): prevent job state flip-flopping between 'queued' and 'processing' (#64098)
Fixes SPLF-170

This fixes a bug in our status reporting where we would report a job as
"queued" although it should be reported as "processing". In some cases
this led to the state flip-flopping between "queued" and "processing".

This happens, for example, if a worker has just finished a repo-rev job
(and set it to completed) but hasn't yet dequeued a new job. In this
brief period we reported the entire search job as queued.

Test plan:
- updated unit test
- manual testing

I ran the following search job which previously exhibited this behavior
and verified that now it doesn't.
```
context:global r:^github\.com/sourcegraph/sourcegraph$@*refs/heads/* ghp_.+ patterntype:regexp
```