chore(search_jobs): add janitor job
This adds a background job to Search Jobs that periodically scans for
finished jobs to aggregate the status, upload logs, and clean up the
tables. This drastically reduces the size of the tables and improves
performance of the API.

For example, with this change a finished search job on .com only has 1
entry in the database, whereas before it could have several milions.

Test plan:
- new unit tests
- manual testing

 I ran a couple of search jobs locally and checked that
 - logs are uploaded to `blobstore-go/buckets/search-jobs`
 - repo jobs are deleted from `exhaustive_repo_jobs`
 - logs are served from the blobstore after the janitor ran