Relax modelconfig name restrictions (#64161)
This PR greatly relaxes the naming restrictions for resource IDs when
using the new `modelconfig` LLM model configuration system.

The original intent was to set a high bar for how we would allow users
to refer to LLM providers, API versions, or models. So that we could
have use those names in URLs and/or display them if no human-friendly
name was supplied. e.g. `opanai::v1::gpt-3.5-turbo`.

However, in practice, this was just a bit too optimistic. We already had
carved out an exception for "API versions" to be less strict, since
there were some cases where we saw those included slashes and things.
And one of our trusted testers immediately ran into problems because the
model name had an `@` in it. (See the linked issue.)

In order to avoid perpetually trying to work around LLM providers and
whatever IDs they use for things, it seemed preferable to just relax our
rules.

With this PR, we will constraint Provider, API Versions, and Model IDs
to only be URL-safe(*). So previously characters like `[]@/+;!` were
prohibited, they are now allowed.) However, we still restrict names to
only include `a-zA-Z` and no unicode characters, and curly braces are
not allowed. (But `[]` and `()` are.

> Pedantically, we do not want to say they are _exactly_ URL safe.
Because we don't perform any sort of escaping (e.g. "%20" -> " ", etc.)
Nor do we allow the `:` which _is_ allowed by the RFC, since we are
using that to delimit parts of the model ref.

Anyways, this should provide enough flexibility to avoid most (hopefully
all) problems users would encounter. While still enforcing some type of
standards with regard to how models are defined.

Fixes [PRIME-451 : invalid ModelID (claude-3-5-sonnet@20240620) after
upgrading to
5.5.2463](https://linear.app/sourcegraph/issue/PRIME-451/invalid-modelid-claude-3-5-sonnet20240620-after-upgrading-to-552463).)

## Test plan

Added tests.

## Changelog

NA.