fix autocomplete regression in `main` ("various" provider issue)
This fixes the current regression of autocomplete in `main`, which occurs
both when (and sometimes when not) using the new `modelConfiguration`
site config option.

In 3cb1c4528dd48530157d5186caf838799886256a (which lives only in `main`,
not in any official releases - except those images that went to Self-hosted
models EAP customers.) we had a regression where `CodyLLMConfiguration.provider`
would return `"various"`.

This change makes it so that we never return `"various"`, and instead only
return provider values that would be accepted by the autocomplete provider
tests of the client (`create-provider.test.ts`)

This change was based purely on evaluating what the client actually does
with this data, see the code comment (a story / explanation in itself)
for details.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>