feat(appliance): change update endpoint based on env var
when APPLIANCE_MANAGED is set to true the site-admin page points to
http://localhost:8889
When it is set to false, nothing changes.

Mostly had to track down all the places in the config / types get set.
The go backend generates these. The integration test defaults to false
(though we should maybe have another one?)