chore(ci): post instructions for PR sections in a comment (#64176)
Follow-up to https://github.com/sourcegraph/devx-support/issues/1130 

Moving the doc links to a comment is merely a few lines of GitHub
Script, so it's worth doing it for better merge commit messages.

## Test plan

This PR itself was the test, as you can see below the comments that
resulted from this new GH workflow. Btw, the second one is the one
you'll see in the code here. I went for a oneliner to avoid cluttering
the UI.