soap: Fix auto-redirect to IdP when SOAP is enabled
When exactly 1 auth provider is configured, Sourcegraph redirects users automatically to the IdP to speed up the sign-in process, so that users don't have to make an extra click to select the one and only sign-in provider.

SOAP is a special case though because it is hidden by default, but enabled on all cloud instances. That caused this auto redirect to never fire for Cloud, since there are technically two auth providers.

This PR fixes it by checking for the sourcegraph-operator query parameter which tells the UI to show the magic SOAP auth provider in the list.

Closes SRC-500

Test plan: Tested on a cloud instance that indeed there is no auto redirect. Then tested locally with SOAP configured that auto redirects happen after this PR, and that there is no auto redirect when the ?sourcegraph-operator query parameter is set.