chore(search): update logging of search durations
This updates and simplifies our logging as follows:
- Prefer to log result types, fall back to pattern type.
- Don't guess repo and file results types if not explicitly specified.
- Allow "AND" and "OR" operators.

Notes:
- I left "New" and "legacy" logging in place. I am not sure how we use
the target stores right now so I wanted to keep this as-is for now.
However we should update this.
- The previous and current logic translates `type:file` queries to
pattern types. This makes sense from the perspective of how things are
implemented in the backend, but it might be confusing for the consumer of
the reporting.

Test plan:
updated unit test