fix(search_jobs): progress reporting
Relates to #64186

With this PR we only show `83 out of 120 tasks` if the search job is
currently processing. In all other states, we don't show this stats.
This is a consequence of the janitor job. After aggregation, this data
is not available anymore.

I remove an unncessary restriction on the download of logs and results.

Test plan:
I ran a search job locally and confirmed that the progress message is
only visible while the job is processing and that logs and downloads are
always available.